# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tandatanganin Digital Signature API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  signed: "signed",
  rejected: "rejected",
} as const;

export interface Document {
  id: number;
  title: string;
  description?: string;
  fileName: string;
  fileSize: number;
  status: DocumentStatus;
  signerName: string;
  signerEmail: string;
  signedAt?: string | null;
  signatureData?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDocumentBody {
  title: string;
  description?: string;
  fileName: string;
  fileSize: number;
  signerName: string;
  signerEmail: string;
}

export type UpdateDocumentBodyStatus =
  (typeof UpdateDocumentBodyStatus)[keyof typeof UpdateDocumentBodyStatus];

export const UpdateDocumentBodyStatus = {
  pending: "pending",
  signed: "signed",
  rejected: "rejected",
} as const;

export interface UpdateDocumentBody {
  title?: string;
  description?: string;
  status?: UpdateDocumentBodyStatus;
  signerName?: string;
  signerEmail?: string;
}

export interface SignDocumentBody {
  signatureData: string;
  signatureId?: number | null;
}

export type SignatureType = (typeof SignatureType)[keyof typeof SignatureType];

export const SignatureType = {
  drawn: "drawn",
  typed: "typed",
  uploaded: "uploaded",
} as const;

export interface Signature {
  id: number;
  name: string;
  signatureData: string;
  type: SignatureType;
  createdAt: string;
}

export type CreateSignatureBodyType =
  (typeof CreateSignatureBodyType)[keyof typeof CreateSignatureBodyType];

export const CreateSignatureBodyType = {
  drawn: "drawn",
  typed: "typed",
  uploaded: "uploaded",
} as const;

export interface CreateSignatureBody {
  name: string;
  signatureData: string;
  type: CreateSignatureBodyType;
}

export interface DashboardSummary {
  totalDocuments: number;
  pendingDocuments: number;
  signedDocuments: number;
  rejectedDocuments: number;
  signaturesThisMonth: number;
  totalSignatures: number;
}

export type ActivityItemAction =
  (typeof ActivityItemAction)[keyof typeof ActivityItemAction];

export const ActivityItemAction = {
  uploaded: "uploaded",
  signed: "signed",
  rejected: "rejected",
  viewed: "viewed",
} as const;

export interface ActivityItem {
  id: number;
  documentId: number;
  documentTitle: string;
  action: ActivityItemAction;
  signerName: string;
  timestamp: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  superadmin: "superadmin",
  admin: "admin",
  user: "user",
  approver: "approver",
} as const;

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: AuthUserRole;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  superadmin: "superadmin",
  admin: "admin",
  user: "user",
  approver: "approver",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  companyName?: string | null;
  division?: string | null;
  role: UserProfileRole;
  isActive: boolean;
  hasGoogleSSO: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
  approver: "approver",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  phone: string;
  companyName?: string;
  division?: string;
  role: CreateUserBodyRole;
  password?: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  user: "user",
  approver: "approver",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  phone?: string;
  companyName?: string;
  division?: string;
  role?: UpdateUserBodyRole;
  isActive?: boolean;
  password?: string;
}

export type Logout200 = {
  ok: boolean;
};

export type ListDocumentsParams = {
  status?: ListDocumentsStatus;
  search?: string;
};

export type ListDocumentsStatus =
  (typeof ListDocumentsStatus)[keyof typeof ListDocumentsStatus];

export const ListDocumentsStatus = {
  pending: "pending",
  signed: "signed",
  rejected: "rejected",
  all: "all",
} as const;